handle all style

+++ app/components/Roles/RolesCard.tsx
@@ -7,20 +7,49 @@ import { useEffect, useState } from "react";
 import { BsTrash2 } from "react-icons/bs";
 import { useDispatch, useSelector } from "react-redux";
 import { toast } from "react-toastify";
+import Swal from "sweetalert2";
 import RolesDetailsCard from "./RolesDetailsCard";
+
+const swalWithBootstrapButtons = Swal.mixin({
+  customClass: {
+    confirmButton: "bg-primary text-white mx-2 px-4 py-2 rounded-md",
+    cancelButton: "bg-red-500 text-white mx-2 px-4 py-2 rounded-md",
+  },
+});
 
 const RolesCard = ({ role, isOpen, openRole, closeRole }: { role: Role; isOpen: boolean; openRole: () => void; closeRole: () => void }) => {
   const dispatch = useDispatch<AppDispatch>();
   const { singleRole } = useSelector((state: RootState) => state.roles);
   const [permissions, setPermissions] = useState<Permission[]>([]);
 
-  const handleDelete = async () => {
-    try {
-      await dispatch(deleteRole(role.id)).unwrap();
-      toast.success("Role deleted successfully");
-    } catch (err) {
-      toast.error(err as string);
-    }
+  const handleDelete = () => {
+    swalWithBootstrapButtons
+      .fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Yes, delete it!",
+        cancelButtonText: "No, cancel!",
+        reverseButtons: true,
+      })
+      .then(async (result) => {
+        if (result.isConfirmed) {
+          try {
+            await dispatch(deleteRole(role.id)).unwrap();
+
+            swalWithBootstrapButtons.fire({
+              title: "Deleted!",
+              text: "Your role has been deleted.",
+              icon: "success",
+            });
+
+            toast.success("Role deleted successfully");
+          } catch (err) {
+            toast.error(err as string);
+          }
+        }
+      });
   };
 
   useEffect(() => {
@@ -73,7 +102,7 @@ const RolesCard = ({ role, isOpen, openRole, closeRole }: { role: Role; isOpen: 
 
             <motion.div
               key="modal"
-              className="fixed top-1/2 left-1/2 z-50 bg-white p-8 rounded-md h-[90%] w-[600px] overflow-hidden overflow-y-scroll "
+              className="fixed top-1/2 left-1/2 z-50 bg-white p-8 rounded-md h-[90%] w-[90%] max-w-[600px] overflow-hidden overflow-y-scroll"
               initial={{ opacity: 0, scale: 0.9, y: "-50%", x: "-50%" }}
               animate={{ opacity: 1, scale: 1, y: "-50%", x: "-50%" }}
               exit={{ opacity: 0, scale: 0.9 }}
